add secret test
