+++ common/db/secret.test.ts
@@ -3,7 +3,7 @@
 /* eslint-disable functional/no-conditional-statement */
 /* eslint-disable functional/no-class */
 import test from 'ava'
-import { writer, reader, Secret } from './db'
+import { writer, reader, Secret } from './secret'
 import { CosmosClient } from '@azure/cosmos'
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
