+++ common/db/secret.ts
@@ -1,5 +1,6 @@
 import { CosmosClient, ItemResponse } from '@azure/cosmos'
-import { createDBInstance } from './common'
+import { always } from 'ramda'
+import { createDBInstance } from './/common'
 
 export type Secret = {
 	readonly id: string
@@ -10,7 +11,17 @@ const SECRETS = {
 	database: 'Authentication',
 	container: 'Secrets',
 }
-export const readSecret = (client: typeof CosmosClient) => async (
+
+export const writer = (client: typeof CosmosClient) => async (
+	data: Secret
+): Promise<ItemResponse<Secret>> => {
+	const container = await createDBInstance(client, SECRETS, process.env)
+	return container.items
+		.create(data)
+		.catch(always(((id) => container.item(id, id).replace(data))(data.id)))
+}
+
+export const reader = (client: typeof CosmosClient) => async (
 	id: string
 ): Promise<ItemResponse<Secret>> => {
 	const container = await createDBInstance(client, SECRETS, process.env)
