+++ event/getSecret/getSecret.ts
@@ -1,5 +1,5 @@
 import { CosmosClient, ItemResponse } from '@azure/cosmos'
-import { readSecret, Secret } from './../../common/db/secret'
+import { reader, Secret } from './../../common/db/secret'
 
 export type oracleArgInfo = {
 	readonly secret: ItemResponse<Secret>
@@ -7,7 +7,7 @@ export type oracleArgInfo = {
 }
 
 export const getSecret = async (json: any): Promise<oracleArgInfo> => {
-	const secret = await readSecret(CosmosClient)(json.s)
+	const secret = await reader(CosmosClient)(json.s)
 	return {
 		secret,
 		json,
