+++ sign/index.test.ts
@@ -2,7 +2,7 @@ import test from 'ava'
 import { HttpRequest, Context } from '@azure/functions'
 import { stub } from 'sinon'
 import * as recover from './recover/recover'
-import * as db from './db/db'
+import * as db from './../common/db/secret'
 import * as importAuthorizer from './importAuthorizer/importAuthorizer'
 import { sign as fakeSignature } from 'jsonwebtoken'
 import sign from './index'
