+++ sign/index.ts
@@ -1,7 +1,7 @@
 import { AzureFunction, Context, HttpRequest } from '@azure/functions'
 import { recover } from './recover/recover'
 import { publicSignature as pubSig } from './publicSignature/publicSignature'
-import { writer } from './db/db'
+import { writer } from './../common/db/secret'
 import { CosmosClient } from '@azure/cosmos'
 import { importAuthorizer } from './importAuthorizer/importAuthorizer'
